requirements.txt
transformers
plotly
dash
pandas
numpy
python-dotenv
torch
wordcloud
nltk